Setup NationBuilder register and login via NationBuilder
Shifted around code to try and make sense of this mess.

I really need to be more disciplined and commit more frequently with
comments, which will make it easier to remember what I did.  I'm having
trouble remembering now.  Suffice to say, I have a Twitter log in
feature.  Can register and log in with NationBuilder.

I am going to have to remove the log in with NationBuilder, as everytime
it's done, NationBuilder will register a new instance of gCanvass on
your profile.  They don't seem to be using good OAuth 2.0 practices of
giving tokens time limits, with the only advantage being that the token
is for the specific user instance it's logging in with.  This saves us
from manually creating a token for them everytime they want to use it.
I'm still not sure if this is going to work the way I want it to, but
I've got to start somewhere.