Modified registration process
Registration goes. go to NationBuilder and login there. Register this
app with your profile, and get a access token. Get user info from
NationBuilder and use this to create a user. Store token along with the
created user associated with that token. Redirect to app, which should
display final part of registration, where a password is chosen.  Store
this password and redirect to app again, completing the process, thus
logging them in according to app.  This should allow them to get the
lists of names, which to choose from.